[compiler]  Splice constraints into .verilog.json [pnr] Move pnr constraint writer to pnr
